moved create_materials()
